+++ src/evdev.h
@@ -359,18 +359,18 @@ evdev_notify_resumed_device(struct evdev_device *device);
 
 void
 evdev_keyboard_notify_key(struct evdev_device *device,
-			  uint32_t time,
+			  uint64_t time,
 			  int key,
 			  enum libinput_key_state state);
 
 void
 evdev_pointer_notify_button(struct evdev_device *device,
-			    uint32_t time,
+			    uint64_t time,
 			    int button,
 			    enum libinput_button_state state);
 void
 evdev_pointer_notify_physical_button(struct evdev_device *device,
-				     uint32_t time,
+				     uint64_t time,
 				     int button,
 				     enum libinput_button_state state);
